code improved, animation problem solved
