+++ tastytoast/src/main/java/com/sdsmdg/tastytoast/ErrorToastView.java
@@ -42,13 +42,14 @@ public class ErrorToastView extends View {
     }
 
     @Override
-    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-        super.onMeasure(widthMeasureSpec, heightMeasureSpec);
-        initPaint();
-        initRect();
+    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
+        super.onSizeChanged(w, h, oldw, oldh);
         mWidth = getMeasuredWidth();
         mPadding = dip2px(10);
         mEyeWidth = dip2px(3);
+        initPaint();
+        initRect();
+        startAnim();
     }
 
     private void initPaint() {
