+++ tastytoast/src/main/java/com/sdsmdg/tastytoast/InfoToastView.java
@@ -42,14 +42,15 @@ public class InfoToastView extends View {
     }
 
     @Override
-    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-        super.onMeasure(widthMeasureSpec, heightMeasureSpec);
-        initPaint();
-        initRect();
+    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
+        super.onSizeChanged(w, h, oldw, oldh);
         mWidth = getMeasuredWidth();
         mPadding = dip2px(10);
         mEyeWidth = dip2px(3);
         endPoint = mPadding;
+        initPaint();
+        initRect();
+        startAnim();
     }
 
     private void initPaint() {
