+++ tastytoast/src/main/java/com/sdsmdg/tastytoast/SuccessToastView.java
@@ -39,13 +39,15 @@ public class SuccessToastView extends View {
     }
 
     @Override
-    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-        super.onMeasure(widthMeasureSpec, heightMeasureSpec);
-        initPaint();
-        initRect();
+    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
+        super.onSizeChanged(w, h, oldw, oldh);
+
         mWidth = getMeasuredWidth();
         mPadding = dip2px(10);
         mEyeWidth = dip2px(3);
+        initPaint();
+        initRect();
+        startAnim();
     }
 
     private void initPaint() {
