+++ tastytoast/src/main/java/com/sdsmdg/tastytoast/TastyToast.java
@@ -7,11 +7,6 @@ import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.facebook.rebound.SimpleSpringListener;
-import com.facebook.rebound.Spring;
-import com.facebook.rebound.SpringConfig;
-import com.facebook.rebound.SpringSystem;
-
 /**
  * Created by rahul on 28/7/16.
  */
@@ -27,13 +22,6 @@ public class TastyToast {
     public static final int DEFAULT = 5;
     public static final int CONFUSING = 6;
 
-    static SuccessToastView successToastView;
-    static WarningToastView warningToastView;
-    static ErrorToastView errorToastView;
-    static InfoToastView infoToastView;
-    static DefaultToastView defaultToastView;
-    static ConfusingToastView confusingToastView;
-
     public static Toast makeText(Context context, String msg, int length, int type) {
 
         Toast toast = new Toast(context);
@@ -42,13 +30,13 @@ public class TastyToast {
         switch (type) {
             case 1: {
                 View layout = LayoutInflater.from(context).inflate(R.layout.success_toast_layout, null, false);
+
                 TextView text = (TextView) layout.findViewById(R.id.toastMessage);
                 text.setText(msg);
-                successToastView = (SuccessToastView) layout.findViewById(R.id.successView);
-                successToastView.startAnim();
                 text.setBackgroundResource(R.drawable.success_toast);
                 text.setTextColor(Color.parseColor("#FFFFFF"));
                 toast.setView(layout);
+                layout.findViewById(R.id.successView);
                 break;
             }
             case 2: {
@@ -56,39 +44,10 @@ public class TastyToast {
 
                 TextView text = (TextView) layout.findViewById(R.id.toastMessage);
                 text.setText(msg);
-
-                warningToastView = (WarningToastView) layout.findViewById(R.id.warningView);
-                SpringSystem springSystem = SpringSystem.create();
-                final Spring spring = springSystem.createSpring();
-                spring.setCurrentValue(1.8);
-                SpringConfig config = new SpringConfig(40, 5);
-                spring.setSpringConfig(config);
-                spring.addListener(new SimpleSpringListener() {
-
-                    @Override
-                    public void onSpringUpdate(Spring spring) {
-                        float value = (float) spring.getCurrentValue();
-                        float scale = (float) (0.9f - (value * 0.5f));
-
-                        warningToastView.setScaleX(scale);
-                        warningToastView.setScaleY(scale);
-                    }
-                });
-                Thread t = new Thread(new Runnable() {
-                    @Override
-                    public void run() {
-                        try {
-                            Thread.sleep(500);
-                        } catch (InterruptedException e) {
-                        }
-                        spring.setEndValue(0.4f);
-                    }
-                });
-
-                t.start();
                 text.setBackgroundResource(R.drawable.warning_toast);
                 text.setTextColor(Color.parseColor("#FFFFFF"));
                 toast.setView(layout);
+                layout.findViewById(R.id.warningView);
                 break;
             }
             case 3: {
@@ -96,11 +55,10 @@ public class TastyToast {
 
                 TextView text = (TextView) layout.findViewById(R.id.toastMessage);
                 text.setText(msg);
-                errorToastView = (ErrorToastView) layout.findViewById(R.id.errorView);
-                errorToastView.startAnim();
                 text.setBackgroundResource(R.drawable.error_toast);
                 text.setTextColor(Color.parseColor("#FFFFFF"));
                 toast.setView(layout);
+                layout.findViewById(R.id.errorView);
                 break;
             }
             case 4: {
@@ -108,11 +66,10 @@ public class TastyToast {
 
                 TextView text = (TextView) layout.findViewById(R.id.toastMessage);
                 text.setText(msg);
-                infoToastView = (InfoToastView) layout.findViewById(R.id.infoView);
-                infoToastView.startAnim();
                 text.setBackgroundResource(R.drawable.info_toast);
                 text.setTextColor(Color.parseColor("#FFFFFF"));
                 toast.setView(layout);
+                layout.findViewById(R.id.infoView);
                 break;
             }
             case 5: {
@@ -120,11 +77,10 @@ public class TastyToast {
 
                 TextView text = (TextView) layout.findViewById(R.id.toastMessage);
                 text.setText(msg);
-                defaultToastView = (DefaultToastView) layout.findViewById(R.id.defaultView);
-                defaultToastView.startAnim();
                 text.setBackgroundResource(R.drawable.default_toast);
                 text.setTextColor(Color.parseColor("#FFFFFF"));
                 toast.setView(layout);
+                layout.findViewById(R.id.defaultView);
                 break;
             }
             case 6: {
@@ -132,11 +88,10 @@ public class TastyToast {
 
                 TextView text = (TextView) layout.findViewById(R.id.toastMessage);
                 text.setText(msg);
-                confusingToastView = (ConfusingToastView) layout.findViewById(R.id.confusingView);
-                confusingToastView.startAnim();
                 text.setBackgroundResource(R.drawable.confusing_toast);
                 text.setTextColor(Color.parseColor("#FFFFFF"));
                 toast.setView(layout);
+                layout.findViewById(R.id.confusingView);
                 break;
             }
         }
